pyproject.toml
[project]
name = "evidence-mcp"
version = "0.1.0"
description = "MCP server for Evidence AI Assistant"
readme = "README.md"
authors = [
    { name = "Jason Ho" }
]
keywords = ["mcp", "evidence", "ai", "assistant", "llm", "dashboard", "reporting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-frontmatter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/jaho5/evidence-mcp"
Repository = "https://github.com/jaho5/evidence-mcp"
Issues = "https://github.com/jaho5/evidence-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
evidence-mcp = "evidence_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/evidence_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
